Move helpers and selectors into same subfolder as integrations integration tests
