+++ ui/apps/platform/cypress/integration/integrations/clusterInitBundles.test.js
@@ -1,10 +1,10 @@
-import { selectors } from '../../constants/IntegrationsPage';
 import withAuth from '../../helpers/basicAuth';
 import {
     getHelperElementByLabel,
     getInputByLabel,
     generateNameWithRandomString,
 } from '../../helpers/formHelpers';
+
 import {
     assertIntegrationsTable,
     clickCreateNewIntegrationInTable,
@@ -12,7 +12,8 @@ import {
     generateCreatedAuthProvidersIntegrationInForm,
     revokeAuthProvidersIntegrationInTable,
     visitIntegrationsTable,
-} from '../../helpers/integrations';
+} from './integrations.helpers';
+import { selectors } from './integrations.selectors';
 
 // Page address segments are the source of truth for integrationSource and integrationType.
 const integrationSource = 'authProviders';
