+++ ui/apps/platform/cypress/integration/integrations/integrations.selectors.js
